Style fixes
